% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_postcodes.R
\name{find_postcodes}
\alias{find_postcodes}
\title{find_postcodes}
\usage{
find_postcodes(suburbs)
}
\arguments{
\item{suburbs}{A character vector. The name of one, or more suburbs}
}
\value{
A list containing a map of the ilocs, and suburb and iloc demographics
}
\description{
takes in suburb names, outputs postcodes
}
