% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_suburbs.R
\name{find_suburbs}
\alias{find_suburbs}
\title{find_suburbs}
\usage{
find_suburbs(postcodes = NULL)
}
\arguments{
\item{postcodes}{A character vector. The name of one, or more postcodes}
}
\value{
A list containing a map of the ilocs, and suburb and iloc demographics
}
\description{
takes in suburb names, outputs postcodes
}
