% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iloc_map.R
\name{iarea_demog}
\alias{iarea_demog}
\title{iarea_demog}
\usage{
iarea_demog(..., iarea = "IARE401016")
}
\arguments{
\item{iarea}{indigenous area}

\item{variables}{variables to search for}
}
\description{
iarea demographics deatils
}
