% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iloc_find.R
\name{iloc_find}
\alias{iloc_find}
\title{iloc_find}
\usage{
iloc_find(x)
}
\arguments{
\item{x}{A character vector. The start of the name of one, or more ILOC regions}
}
\value{
a tibble containing the names of all ILOCs
}
\description{
Allows a user to input the start of an iloc name. All matching names are returned.
}
