% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iloc_map.R
\name{iloc_map}
\alias{iloc_map}
\title{iloc_map}
\usage{
iloc_map(iloc, maptype = "watercolor", zoom = 12, postcodes = T,
  gap = 0.07, ...)
}
\arguments{
\item{iloc}{A charcter vector. The name of one, or more ILOC regions}

\item{maptype}{see get_stamenmap}

\item{zoom}{numeric. Defines the resolution of the map. Lower resolution loads faster}

\item{postcodes}{bool. If True, postcodes will be output for suburbs.}

\item{gap}{numeric. Adds this to long and lat. Larger numbers result in larger displayed area}

\item{...}{extra arguments can be supplied to get_stamenmap.}
}
\value{
A list containing a map of the ilocs, and suburb and iloc demographics
}
\description{
iloc_map takes in a vector of ILOC names, and returns a map and demographic information.
}
