% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iloc_sample.R
\name{iloc_sample}
\alias{iloc_sample}
\title{iloc_sample}
\usage{
iloc_sample(x, n = NULL, p = NULL)
}
\arguments{
\item{x}{An iloc_demog table.}

\item{n}{the sample to adjust based on. tibble}
}
\value{
A tibble
}
\description{
iloc_sample takes in a matrix or dataframe. It calculates the element wise proportion, and then returns the table based on a different total.
}
