% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/just_map.r
\name{map_only}
\alias{map_only}
\title{map_only}
\usage{
map_only(iloc = "Port Adelaide", maptype = "watercolor", zoom = 5,
  gap = 0.07, left_adjust = 0, bottom_adjust = 0, right_adjust = 0,
  top_adjust = 0, ...)
}
\arguments{
\item{iloc}{name of iloc}

\item{maptype}{what map type do you want?}

\item{zoom}{smaller number is lower resolution}

\item{gap}{area around bbox to apply padding}

\item{left_adjust}{change left side}

\item{bottom_adjust}{change bottom}

\item{right_adjust}{you get it}

\item{top_adjust}{as above.}
}
\description{
Just grabs a map based on provided ilocs
}
