% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iloc_map.R
\name{iloc_map}
\alias{iloc_map}
\title{iloc_map}
\usage{
iloc_map(iloc, maptype = "watercolor", zoom = 12, postcodes = T,
  gap = 0.07, map = T, label = T, left_adjust = 0,
  bottom_adjust = 0, right_adjust = 0, top_adjust = 0, ...)
}
\arguments{
\item{iloc}{A charcter vector. The name of one, or more ILOC regions}

\item{maptype}{see get_stamenmap}

\item{zoom}{numeric. Defines the resolution of the map. Lower resolution loads faster}

\item{postcodes}{bool. If True, postcodes will be output for suburbs.}

\item{gap}{numeric. Adds this to long and lat. Larger numbers result in larger displayed area}

\item{map}{bool. If TRUE, this function will output a map.}

\item{label}{bool. If TRUE \(default\) the largest suburbs will be labelled.}

\item{left_adjust}{a numeric.}

\item{bottom_adjust}{a numeric.}

\item{right_adjust}{a numeric.}

\item{top_adjust}{a numeric.}

\item{...}{extra arguments can be supplied to get_stamenmap.}
}
\value{
A list containing a map of the ilocs, and suburb and iloc demographics
}
\description{
iloc_map takes in a vector of ILOC names, and returns a map and demographic information.
}
